Update CREATE TABLE sales.sql

+++ CREATE TABLE sales.sql
@@ -4,7 +4,8 @@ CREATE TABLE sales (
 	release_date text,
 	title text,
 	genre text,
-	international_box_office integer,
+	international_box_office BIGINT,
 	domestic_box_office text,
-	worldwide_box_office integer
+	worldwide_box_office BIGINT
+
 );
